+++ CentroLink/CentroLink/PromoTicketSetupModule/DatabaseEntities/PromoTicketSetupEventTypes.cs
@@ -6,6 +6,8 @@
         PromoTicketAdded = 1118,
         PromoTicketAddedFailed = 1119,
         PromoTicketModified = 1120,
-        PromoTicketModifiedFailed = 1121
+        PromoTicketModifiedFailed = 1121,
+        PromoTicketDeleted=1122,
+        PromoTicketDeletedFailed=1123
     }
 }