+++ CentroLink/CentroLink/PromoTicketSetupModule/Services/IPromoTicketSetupService.cs
@@ -10,5 +10,7 @@ namespace CentroLink.PromoTicketSetupModule.Services
         PromoTicketModel GetPromoTicketEditModel(int promoScheduleId);
         void CreatePromoTicket(PromoTicketModel promoTicket);
         void UpdatePromoTicket(PromoTicketModel promoTicket);
+        void DeletePromoTicket(int promoScheduleId);
+        void StopScheduleItem(int promoScheduleId);
     }
 }